Algebra

Description

Implementation of relational algebra and other operators over Relvars as defined in C.J. Date, "An Introduction to Database Systems" Eighth Edition, chapter 7 ('Relational Algebra') .

Documentation

module Relvar

Types

data OpElem

Summarize operators.

Constructors

Sum

Count
Count elements.

Avg
Arithmetic Average.

Max
Return the maximum element.

Min
Return the minimum element.

Sdev
Standard deviation.

Var
Variance.

Algebra operators

union :: Relvar -> Relvar -> Relvar

Union operator. The two relations' attributes must be equal.

intersection :: Relvar -> Relvar -> Relvar

Intersection operator The two relations' attributes must be equal (see match).

minus :: Relvar -> Relvar -> Relvar

Difference operator. The two relations' attributes must be equal (see match).

times :: Relvar -> Relvar -> Relvar

Product operator. The two relations' attributes must be disjoint.

restrict :: Relvar -> ([Elem] -> Bool) -> [String] -> Relvar

Filter a relation's rows given a ([Elem] -> Bool) function and a list of attribute names.

project :: Relvar -> [String] -> Relvar

Projection on a given set of attributes. Only attributes with labels in the provided list are included.

projectaway :: Relvar -> [String] -> Relvar

Projection on a given set of attributes. Attributes with labels matching the list are excluded.

join :: Relvar -> Relvar -> Relvar

Natural join operator. The two relations' attributes must not be disjoint. The resulting

Other operators

semiJoin :: Relvar -> Relvar -> Relvar

Similar to the join operator projected over the attributes of the first relation. Loosely speaking the results of semiJoin r1 r2 are the rows of r1 which have a counterpart in r2.

r1 `semiJoin` r2 == project (r1 `join` r2) (labels r1)

semiMinus :: Relvar -> Relvar -> Relvar

Similar to the minus operator projected over the attributes of the first relation. Loosely speaking the results of semiMinus r1 r2 are the rows of r1 which do not have a counterpart in r2.

r1 `semiMinus` r2 == project (r1 `minus` r2) (labels r1)

r1 `semiMinus` r2 == r1 `minus` (r1 `semiJoin` r2)

extend :: Relvar -> ([Elem] -> Elem) -> [String] -> String -> Relvar

Add a derived column to an existing relation. Each element in the new column is the result of a computation of some other columns.

summarize :: Relvar -> Relvar -> [OpElem] -> [String] -> [String] -> Relvar

summarize r per ops inlbls outlbls computes ops over inlbls columns into outlbls column.

group :: Relvar -> [String] -> String -> Relvar

group r grplbls reslbl returns a relation with all of r's attributes, excluding grplbls columns, plus reslbl column. The elements in reslbl column are of type Relvar and the result of grouping of grplbls columns over the non-grplbls of r.

degree (group r grplbls reslbl) == degree r - (length grplbls) + 1

ungroup :: Relvar -> String -> Relvar

group inverse operator.