Algebra.Function

Description

Boilerplate code to translate a->b functions into corresponding 'Relvar.Elem'->'Relvar.Elem' functions (see liftElem) and also into [Elem]->Elem functions for the function extend.

Documentation

module Relvar

Classes

class (Elementable a, Elementable b) => LiftElem a b where

Lifting functions into the "Elem domain" Use this class to translate f :: a -> b into the corresponding (Elem -> Elem) function

Minimal complete definition

Nothing

Methods

liftLeft' :: (a -> b) -> Elem -> Maybe b

liftLeft :: (a -> b) -> [Elem] -> Maybe b

liftEl' :: (a -> b) -> Elem -> Elem

liftEl :: (a -> b) -> [Elem] -> Elem

class (Elementable a1, Elementable a2, Elementable b) => LiftElem2 a1 a2 b where

Lifting functions into the "Elem domain" Use this class to translate f :: a1 -> a2 -> b into the corresponding (Elem -> Elem -> Elem) function

Minimal complete definition

Nothing

Methods

liftLeft2' :: (a1 -> a2 -> b) -> Elem -> Elem -> Maybe b

liftLeft2 :: (a1 -> a2 -> b) -> [Elem] -> Maybe b

liftEl2' :: (a1 -> a2 -> b) -> Elem -> Elem -> Elem

liftEl2 :: (a1 -> a2 -> b) -> [Elem] -> Elem

class (Elementable a1, Elementable a2, Elementable a3, Elementable b) => LiftElem3 a1 a2 a3 b where

Lifting functions into the "Elem domain" Use this class to translate f :: a1 -> a2 -> a3 -> b into the corresponding (Elem -> Elem -> Elem -> Elem) function

Minimal complete definition

Nothing

Methods

liftLeft3' :: (a1 -> a2 -> a3 -> b) -> Elem -> Elem -> Elem -> Maybe b

liftLeft3 :: (a1 -> a2 -> a3 -> b) -> [Elem] -> Maybe b

liftEl3' :: (a1 -> a2 -> a3 -> b) -> Elem -> Elem -> Elem -> Elem

liftEl3 :: (a1 -> a2 -> a3 -> b) -> [Elem] -> Elem

class LiftElem a Bool => BoolFun a where

Lifting (a->Bool) functions to ([Elem]->Bool) functions. This is a convenience function when using restrict function.

Minimal complete definition

Nothing

Methods

liftBoolFun :: (a -> Bool) -> [Elem] -> Bool

class LiftElem2 a1 a2 Bool => BoolFun2 a1 a2 where

Lifting (a1->a2->Bool) functions to ([Elem]->Bool) functions. This is a convenience function when using restrict function.

Minimal complete definition

Nothing

Methods

liftBoolFun2 :: (a1 -> a2 -> Bool) -> [Elem] -> Bool

class LiftElem3 a1 a2 a3 Bool => BoolFun3 a1 a2 a3 where

Lifting (a1->a2->Bool) functions to ([Elem]->Bool) functions. This is a convenience function when using restrict function.

Minimal complete definition

Nothing

Methods

liftBoolFun3 :: (a1 -> a2 -> a3 -> Bool) -> [Elem] -> Bool