LabelledSet

Description

Implementation of operators on "labelled sets", LSets, which are a type synonym for Set (String, a).

Relations are defined by their Attributes and the Set of their Rows which are both defined as LSets :

type Attributes = LSet TypeRep

type Row = LSet Elem

Types

type LSet a = Set (String, a)

Class

class LabelSetOps a where

Any LSet and Relvars are part of this class.

For Relvars, the operators apply to the Relvar's Attributes which are defined as LSet TypeRep.

Methods

compatible :: (Ord a, Show a) => a -> a -> Bool

Check that two LSets or two Relvars are "compatible", e.g. there is no label equality without value equality.

match :: Eq a => a -> a -> Bool

LSets equality. For Relvars, True if the Attributes of two relations are equal.

disjoint :: Ord a => a -> a -> Bool

rename :: a -> [String] -> [String] -> a

labels :: a -> [String]

degree :: a -> Int

Construction

insert :: Ord a => String -> a -> LSet a -> LSet a

delete :: String -> LSet a -> LSet a

Combine

project :: LSet a -> [String] -> LSet a

projectaway :: (Ord a, Show a) => LSet a -> [String] -> LSet a

common :: Ord a => LSet a -> LSet a -> LSet a

union :: Ord a => LSet a -> LSet a -> LSet a

Query

extract :: (Ord a, Show a) => Bool -> LSet a -> [String] -> [a]

extract values provided names. Preserve names order.

values :: LSet a -> [a]

toList :: LSet a -> [(String, a)]

Convert the set to a list of (String,Value) elements.

fromList :: Ord a => [String] -> [a] -> LSet a

fromList' :: Ord a => [(String, a)] -> LSet a

Convert the set to a list of values.

Misc

checkLabels :: Bool -> [String] -> Bool

compatibleLabels :: (Ord a, Show a) => Bool -> LSet a -> [String] -> Bool