Relvar

Description

A simple representation of relation types based on C.J. Date, "An Introduction to Database Systems" Eighth Edition.

The representation of Relvar is based on Sets, where each relation tuple is of the type Row, or a synonym for Set (String, Elem). Elem is an Algebraic Data Type representing each tuple's element.

Re-export

data TypeRep :: *

A concrete representation of a (monomorphic) type. TypeRep supports reasonably efficient equality.

Types

data Elem

Type for Relvar's elements. Elem types are either primitive types or Relvars or array of Elems or tuple of Elems (2 and 3 arguments). One can extend the primitive types by adding to the code's definitions (look for "-- ADD NEW TYPES HERE" comments in code).

Constructors

Nil

B Bool

C Char

S String

T Text

I Int

J Integer

D Double

DD Day

DT UTCTime

BS ByteString

A [Elem]
array of Elems

T2 (Elem, Elem)
pair of Elems

T3 (Elem, Elem, Elem)
triplet of Elems

R Relvar
relation

DW (Down Elem)
convenient for ordering Elems in descending order

type Attributes = LSet TypeRep

Attributes is the type synonym for LSet TypeRep, i.e. for Set (String, TypeRep). The LSet type enforces that there are no blank label, nor any duplicate labels.

type Row = LSet Elem

Row is synonym for LSet Elem, i.e. for Set (String, Elem). The LSet type enforces that there are no blank label, nor any duplicate label.

data Relvar

Relvar is the type for relations. A relation is defined by the set of its attributes (i.e. labels' names and types) and of its tuples, or rows, (i.e. sets of Elems). There is unicity of Attributes labels and no two rows can be identical.

See C.J. Date, "An Introduction to Database Systems" Eighth Edition, chapter 6 ('Relations').

Constructors

Relvar

Fields

attributes :: Attributes

Set of the relation's attributes.

rdata :: Set Row

Set of the relation's tuples (i.e. rows).

data Table

Type for tables. A relation's table output is mostly for representation. Unlike relations, tables can have duplicate rows and column headers. See full difference between relations and tables in C.J. Date, "An Introduction to Database Systems" Eighth Edition, chapter 6 ('Relations').

Constructors

Table

Fields

header :: [String]

tdata :: [[Elem]]

data ColOrder

ADT for selecting sorting in ascending or descending orders.

Constructors

Asc

Desc

Classes

class Elementable a where

Class of any type a which can be reprensented as an Elem.

Methods

toElem :: a -> Elem

Conversion from types to their corresponding Elem ADT type.

fromElem :: Elem -> Maybe a

Construction

relvar :: [String] -> [TypeRep] -> [[Elem]] -> Relvar

Create a relation variable provided its Attributes' labels and types (i.e. column names and types) and its Row elements.

dee :: Relvar

Constant relation with an empty set of attributes and 1-tuple (equal to the empty set) body ([C.J. Date] chap. 6).

dee = Relvar {attributes= Set.empty, rdata = Set.singleton $ Set.empty}.

dum :: Relvar

Constant relation with an empty set of attributes and no-tuple body ([C.J. Date] chap. 6).

dum = Relvar {attributes= Set.empty, rdata = Set.empty}.

rename :: LabelOp a => a -> [String] -> [String] -> a

Replace LSet's labels. For Relvars, rename its Attributes's labels.

Query

Information

degree :: LabelOp a => a -> Int

Return the number of elements in the set. For Relvars, return its Attributes's size (e.g. its number of columns).

card :: Relvar -> Int

Cardinality of a relation (i.e. # of tuples or # of rows).

elems :: Relvar -> [[(String, Elem)]]

Extract all elements.

elems' :: Relvar -> Bool -> [String] -> [[Elem]]

Extract elements given labels.

labels :: LabelOp a => a -> [String]

Return LSet's labels. For Relvars, return its Attributes's labels.

types :: Relvar -> [TypeRep]

Return the all Attributes' types.

types' :: Relvar -> Bool -> [String] -> [TypeRep]

Return Attributes' types for the matching labels.

common :: Relvar -> Relvar -> Attributes

Common attributes

Logical

compatible :: (LabelOp a, Ord a, Show a) => a -> a -> Bool

Check that two LSets or two Relvars are "compatible", e.g. there is no label equality without value equality. For Relvars, True if their Attributes are compatible.

match :: (LabelOp a, Eq a) => a -> a -> Bool

LSets equality. For Relvars, True if the Attributes of two relations are equal.

disjoint :: (LabelOp a, Ord a) => a -> a -> Bool

True if LSets are compatible and disjoint.

Table

table :: Relvar -> Maybe [String] -> Maybe [(String, ColOrder)] -> Maybe [Int] -> Table

Output a table given a set of labels, a row indexing function and a list of row indices.

If lbls is Nothing all relation labels are included (in default order). If the row indices argument is nothing, all rows are included.

Label functions

keepLeft :: Relvar -> [String] -> [String]

Given a relation and some of its labels [ai0, ai1, ... ain], returns the relation's attributes reordered with lftlbls on the left and then the remaining labels from lbls.

keepRight :: Relvar -> [String] -> [String]

Given a relation and some of its labels [ai0, ai1, ... ain], returns the relation's attributes reordered with lftlbls on the left and then the remaining labels from lbls.

Utility functions

typeRep :: Elem -> TypeRep

Mapping from Elem to their corresponding typeRep. Only relevant for basic types and relations.

tyB :: TypeRep

TypeRep for Bool.

tyC :: TypeRep

TypeRep for Char.

tyS :: TypeRep

TypeRep for String.

tyT :: TypeRep

TypeRep for Text.

tyI :: TypeRep

TypeRep for Int.

tyJ :: TypeRep

TypeRep for Integer.

tyD :: TypeRep

TypeRep for Double

tyDD :: TypeRep

TypeRep for Day.

tyDT :: TypeRep

TypeRep for UTCTime.

tyBS :: TypeRep

TypeRep for ByteString.

tyR :: TypeRep

TypeRep for Relvar.

tyA :: TypeRep

TypeRep for [Elem].

tyT2 :: TypeRep

TypeRep for (Elem, Elem).

tyT3 :: TypeRep

TypeRep for (Elem, Elem, Elem).

tyZ :: TypeRep

TypeRep for Nil.