Relvar.CSV

Description

Import and convert csv files of basic types elements (only) into Elems and produce the corresponding relation. Note that duplicate rows will be merged into one.

Documentation

data CSVSettings :: *

Settings for a CSV file. This library is intended to be flexible and offer a way to process the majority of text data files out there.

Constructors

CSVSettings

Fields

csvSep :: !Char

Separator character to be used in between fields

csvQuoteChar :: !(Maybe Char)

Quote character that may sometimes be present around fields. If Nothing is given, the library will never expect quotation even if it is present.

Types

fromCSV :: Ord a => Map a TypeRep -> a -> String -> Elem

Given an appropriate TypeRep mapping, convert a CSV string to Elem.

toRelation :: String -> Char -> [String] -> [TypeRep] -> IO Relvar

toRelation filepath delimiter labels types to corresponding IO (Relvar).