Relvar.LSet

Description

Implementation of convenience operators on "labelled sets", LSets, which are a type synonym for Set (String, a), that is a set of (label, value) pairs. Unicity of labels is required and blank labels are prohibited.

Relations are defined by their Attributes and the Set of their Rows which are both defined as LSets :

type Attributes = LSet TypeRep

type Row = LSet Elem

Types

type LSet a = Set (String, a)

Generic Set of (label, value) elements.

Class

class LabelOp a where

Labels operators typeclass. Any LSet and Relvar are part of this class.

For Relvars, the operators apply to the Relvar's Attributes (defined as LSet TypeRep).

Methods

compatible :: (Ord a, Show a) => a -> a -> Bool

Check that two LSets or two Relvars are "compatible", e.g. there is no label equality without value equality. For Relvars, True if their Attributes are compatible.

match :: Eq a => a -> a -> Bool

LSets equality. For Relvars, True if the Attributes of two relations are equal.

disjoint :: Ord a => a -> a -> Bool

True if LSets are compatible and disjoint.

rename :: a -> [String] -> [String] -> a

Replace LSet's labels. For Relvars, rename its Attributes's labels.

labels :: a -> [String]

Return LSet's labels. For Relvars, return its Attributes's labels.

degree :: a -> Int

Return the number of elements in the set. For Relvars, return its Attributes's size (e.g. its number of columns).

Construction

For completeness, but it might be simpler to use the conversion functions (e.g. fromList). Also see LabelOp class instance function rename.

empty :: LSet a

Empty set

singleton :: String -> a -> LSet a

Singleton set from a label and its corresponding element.

insert :: Ord a => String -> a -> LSet a -> LSet a

Insert a label and corresponding value in an LSet. -- CHECK : 1) that the label does not exist already 2) that the label is not blank

delete :: String -> LSet a -> LSet a

Delete an element in an LSet.

Combine

project :: LSet a -> [String] -> LSet a

Return a LSet of the input LSet elements which labels are included in the provided list.

projectaway :: (Ord a, Show a) => LSet a -> [String] -> LSet a

Return a LSet of the input LSet elements which labels are not included in the provided list.

common :: Ord a => LSet a -> LSet a -> LSet a

Return common labels between the two LSets.

union :: Ord a => LSet a -> LSet a -> LSet a

Union of the two LSets.

Query

Information

Also see LabelOp class instance functions labels and degree.

extract :: (Ord a, Show a) => Bool -> LSet a -> [String] -> [a]

extract values provided labels. Preserve input labels order.

extractWithLabels :: (Ord a, Show a) => Bool -> [String] -> LSet a -> [(String, a)]

extract (label, value) provided labels. Preserve input labels order.

extract2 :: Ord a => [String] -> [(String, a)] -> [(String, a)]

values :: LSet a -> [a]

Return values of a given LSet. Note that the order of the list is determined by the internal Set representation of the LSet input.

Logical

See LabelOp class instance functions match, disjoint and compatible.

Conversion

toList :: LSet a -> [(String, a)]

Convert the set to a list of (label,value) pairs.

fromList :: Ord a => [String] -> [a] -> LSet a

Create a LSet from a list of labels and a corresponding list of values.

Miscellaneous

checkLabels :: Bool -> [String] -> Bool

Return True if no blank labels and, if chckDup flag is True, no duplicate labels are found.

compatibleLabels :: (Ord a, Show a) => Bool -> LSet a -> [String] -> Bool

Return True if all labels are included in the provided LSet and if checkLabels is also True.

Similar to compatible but tests the compatibility of LSet's labels vs the provided list of labels.