Relvar.Pretty

Description

Set of convenience functions to render relations and output them to stdout.

Credit: This module is based on a modified version of Text.PrettyPrint.Boxes in the boxes package. Original version by Brent Yorgey and David Feuer here and here. Modified version by rustydc here.

The original boxes package cannot be used as is, as it does not handle Text type (and therefore does not allow Unicode characters output).

Types

data Params

Box parameters.

Constructors

Params

Fields

col_width_min :: Int

border_top :: Bool

border_bot :: Bool

border_lft :: Bool

border_rgt :: Bool

col_align :: Alignment

data Alignment

Data type for specifying the alignment of boxes.

Constructors

AlignFirst
Align at the top/left.

AlignCenter1
Centered, biased to the top/left.

AlignCenter2
Centered, biased to the bottom/right.

AlignLast
Align at the bottom/right.

View Functions

render' :: Tabular a => a -> Params -> [Int] -> Text

Render a Relvar as a Text, suitable for writing to the screen or a file.

render :: Tabular a => a -> Text

Render a Relvar as a Text. Identical to render' but defParams is used and all rows are shown.

view' :: Tabular a => a -> Params -> [Int] -> IO ()

Output a relation to the screen, given parameters and the number of rows to output.

view :: Tabular a => a -> [Int] -> IO ()

Output a relation to the screen. Identical to view' but with defParams.

viewAll :: Tabular a => a -> IO ()

Output a relation to the screen. Identical to viewAll' but with defParams.

viewAll' :: Tabular a => a -> Params -> IO ()

Output a relation to the screen. Identical to view' but showing all rows.

Misc

defParams :: Params

Default parameters.

col_width_min = 10, border_top = False, border_bot = False, border_lft = False, border_rgt = False, col_align = left